+++ ChangeLog.md
@@ -28,6 +28,7 @@ repository history](https://github.com/ddclient/ddclient/commits/main).
     [#907](https://github.com/ddclient/ddclient/pull/907)
   * Added `spaceship` protocol for Spaceship (https://www.spaceship.com/).
     [#896](https://github.com/ddclient/ddclient/pull/896)
+  * Added `dnspod` protocol for DNSPod / Tencent Cloud DNS (https://www.dnspod.com/).
 
 ## 2026-05-16 v4.0.1-rc.1